\name{fitted.csvy}
\alias{fitted.csvy}

\title{
A S3 Method of the Function FITTED for a CSVY Object
}

\description{
fitted.csvy extracted estimated domain mean from a csvy object.
}

\usage{
\method{fitted}{csvy}(object, ...)
}

\arguments{
  \item{object}{A csvy object.}
  \item{\dots}{Additional parameters.}
}

\value{
Estimated domain means.
}