\name{plotpersp.csvy}
\alias{plotpersp.csvy}

\title{Create a 3D Plot for a CSVY Object}

\description{
This routine is a method of the plotpersp routine in the cgam package. 
}

\value{
plotpersp.csy returns a 3D plot of an object of the csvy class. The \eqn{x} lab and \eqn{y} lab represent a set of non-parametrically modelled predictors used in a csvy formula, and the \eqn{z} lab represents the estimated domain mean surface.
}
