\name{print.csvy}
\alias{print.csvy}

\title{
S3 Method of the Function PRINT for a CSVY Object
}

\description{
prints the formula and design in a csvy object and returns it invisibly (via invisible(x)).
}


\usage{
\method{print}{csvy}(x, ...)
}

\arguments{
  \item{x}{A csvy object.}
  \item{\dots}{Additional parameters.}
}

\value{
No return value, called for side effects.
}


